requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.0

# ML/CV - Latest versions compatible with Python 3.13
torch>=2.6.0
torchvision>=0.21.0
timm>=1.0.0
Pillow>=10.0.0
numpy>=1.26.0

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.0.0
aiofiles>=23.2.1

# Logging & Monitoring
python-json-logger>=2.0.7
